requirements.txt
absl-py==2.0.0
asttokens==2.4.1
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.47.0
gast==0.5.4
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
h5py==3.10.0
idna==3.6
imageio==2.33.1
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.0
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydicom==2.4.4
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.12.9
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
